Add link handling.
For a given URL, parse its HTML and retrieve all links. An actor is
spawned for each link, which will check if it is a link to a wikipedia
article. If it is, a message will be sent to the WikipediaParser
actor. If not, the actor will throw an exception and the parent actor will
terminate the actor.